url-analise-service/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
aiohttp==3.9.1
requests==2.31.0  # Используется для Telegram уведомлений
beautifulsoup4==4.12.2
asyncpg==0.29.0
kafka-python==2.0.2
pydantic==2.5.0
python-multipart==0.0.6
